requirements.txt
# Core dependencies
openai==1.3.0
pydub==0.25.1
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3

# API framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# AWS integration
boto3==1.34.0

# Audio/Video processing
moviepy==1.0.3
Pillow==10.0.0

# Additional utilities for production
python-json-logger==2.0.7 